chore: Define expected package manager (#6119)
The "packageManager" field defines which package
manager is expected to be used when working on the
current project. It can set to any of the supported
package managers, and will ensure that your teams
use the exact same package manager versions without
having to install anything else than Node.js. See
https://nodejs.org/dist/latest-v16.x/docs/api/all.html#all_packages_packagemanager
for more information.

This commit sets the expected package manager in
`workers-sdk` to `pnpm@9.1.3`.