quarantine C3 astro e2e test
C3 astro e2e test started failing in CI for no
obvious reasons since Mon 24 June 2024. We
tried everything, like increasing timeouts,
debugging, reproducing locally (which we can't)
to no avail. We are quite confident that the test
failures are not caused by any changes in wrangler
(we verified this by running the tests against a PR
where they were consistently green, but did fail on
the re-run).

As this is blocking our release, we decided to
quarantine these tests, do the release, while continuing
to look into why the tests are failing.