fixup PR:
Rename validateNodeCompat to getNodeCompatMode which is what the function does

Convert the parameters to optional options for better readability.

Added a `validateConfig` to be able to skip validation (used once in the code base).

The former getNodeCompatMode is now parseNodeCompatibilityFlags and is only responsible to return the compat flags and no more exported.

Improve the documentation