Pre-populate existing image when modifying a deployment
If a user wants to update a deployment to change only the tag of their
image, then they must re-enter the entire image URL with the new tag.
Instead, we can pre-populate the image field with the existing image URL
(including the tag), which means a user can now simply modify the tag
without needing to re-enter the full URL themselves.

For example, if the deployment is currently using image
"hello-world:1.0" and they want to modify it to use "hello-world:1.1",
currently they would need to write the full text "hello-world:1.1". This
can be quite cumbersome if the URL of the image is long (which it often
is). With this change, the existing URL is pre-populated, so the user
needs only to replace "0" with "1".

If a user does not want to modify the image of the deployment then they
can continue to just press "Enter" to keep the existing image as before.