Add default image for cloudchamber create
Add a default image when using "wrangler cloudchamber create". This
image is docker.io/cloudflare/hello-world which is a simple HTTP server
that runs on Cloudflare's container platform.

This commit also augments validation to both the `cloudchamber create`
and `cloudchamber modify` commands to ensure the provided image contains
a tag and adheres to the format in the OCI specification.