cloudchamber/create: always print deployment and placement ID
We currently only print the full deployment ID when the deployment has
an IPv4 address. This commit makes the deployment ID always print, as
well as the placement ID. We also move the printing of the IPv4 address
(if one exists) to the same place as the IPv6 address so that they are
printed together.

There are a few other nit-picky changes as well. I've tried to make
capitalization of labels more consistent (there doesn't appear to be a
rule on when someone should be capitalized and when it shouldn't be, but
I've at least made it more consistent within the subcommands
themselves). I also changed the green "SUCCESS" message to only print
once at the very end of the `cloudchamber create` command. This is
consistent with the flow for creating a Worker.