Update help message for `batch-max-*` pipeline parameters
Closes https://jira.cfdata.org/browse/PIPE-152

The defaults are currently in flux and changing as we determine
what makes sense for the system. To support these changing values,
remove the defaults from the client as they will be set by the API.
The values used to create the pipeline are visible to user when doing
`wrangler pipeline show <name>`.

Create three-chefs-bathe.md