feat: add option to `readConfig()` to specify a list of optional environments
In such cases there is no need for the user to define this environment explicitly,
even if there are other environments defined.

This will be used by the Vite plugin, since `vite dev` command will always set the environment
to "development", and we don't want the user to have to define this explicitly if they don't want to.