Relax the Wrangler peer dependency for the Vite plugin
We only need the Wrangler peer dependency to be at least 3.101.0,
which is when we added support for the redirected config.

By using the peer dependency range version in the playgrounds, we ensure that
we are testing against what will be used in practice. It will also ensure that this
dependency gets bumped when needed, since the playgrounds might start to fail otherwise.