CI: compute the name for artifacts correctly
In the `run-c3-e2e` GH Action we compute various things from the inputs. In particular using `experimental` and `quarantine`.

But it seems that these are converted to strings rather than booleans, so we cannot simple check the truthy state of `inputs.quarantine`, for example since this might have the value `"false"`, which is truthy.