feat(miniflare): Add assets Proxy Worker skeleton in miniflare
This commit implements a very basic Proxy Worker skeleton, and
wires it in the "pipeline" miniflare creates for assets. This
Worker will be incrementally worked on, but for now, the current
implementation will forward all incoming requests to the Router
Worker, thus leaving the current assets behaviour in local dev,
the same.

Fixes DEVX-1641, DEVX-1642