fix: make vitest-pool-workers not crash with scripts that have Workflows (#8208)
* fix: make vitest-pool-workers not crash with scripts that have Workflows

Because `vitest-pool-workers` uses its runner worker as the entrypoint
and not the user provided worker. The user provided worker is loaded
AFAICT though eval (`vitest-pool-workers` enables eval) so we need, at
startup time, to add the required entrypoints and make them dynamically
call the user one.

Not only that, if you try to create a workflow instance it would throw a
`isolated storage` error - I suspect that is this caused by engine still
being active at the end of the test (since workflows are async, not in
the request path, and might take undefinite time to complete).

In order to fix this, I've bypassed the isolated storage check for
workflows internal DOs. I wonder if there might be collisions between
test files if there are any concurrency regarding test files (aka
`singleWorker` is false) - I think that this is fine to do, the point of
this PR is not to make workflows really testable yet but to at least
to make `vitest-pool-workers` **not** crash if a user includes a
workflow in the worker definition.

* chore: add proper readme

* chore: remove isolatedStorage bypass and error out early if isolated storage is used

* chore: add suggestions

* chore: moar test

* Apply suggestions from code review

Co-authored-by: Pete Bacon Darwin <pete@bacondarwin.com>

* Apply suggestions from code review

Co-authored-by: Pete Bacon Darwin <pete@bacondarwin.com>

* stream logs

* no more hanging

* Revert "stream logs"

This reverts commit dfecf634c28b121cb51359e8840e0f352484864c.

---------

Co-authored-by: Pete Bacon Darwin <pete@bacondarwin.com>
Co-authored-by: Samuel Macleod <smacleod@cloudflare.com>