WC-3584 Deny requests to the user worker when over free tier limits
Returns an html page (included as direct module, via esbuild) with a 429
status code, mirroring the current free tier limiting behavior. It also
sets a new field in our analytics so we can observe free tier invocation
denials.

This page can evolve as we need it to, this is just my first pass at the
response.