WC-3584 Move the limited response page to be inlined
There's no loader configured for html files by default in wrangler it
seems. So rather than mess with loaders or a custom build, inlining this
page is easy enough.

This pages copies what fl returns exactly, and there's a few templated
things to reproduce that. Namely: the hostname in the title of the page,
the ip address of the eyeball, and the current date (with some formatting)