Update CONTRIBUTING for hotfixes
It is not necessary to create a changeset for the hotfix, this would only trigger another release of the package during the next normal release process. If such a release is necessary it would be triggered by the existence of a changeset for the feature/fix that needs releasing.

What is important is that the version in package.json is equal to the latest released version of the package. If it is behind then any changesets for that package that should trigger a release may only bump the version to the hotfix release version and so would not actually trigger a release of that package.