Block requests vulnerable to opennext vulnerability (#9635)
Addresses https://github.com/opennextjs/opennextjs-cloudflare/security/advisories/GHSA-rvpw-p7vw-wj3m

This has already been applied in production, this commit keeps the
repository in sync.

Images loaded via the /_next/image?url=<some remote image> were
vulnerable to SSRF, and particularly if the user loaded the image in a
new tab (instead of forcing the browser to treat it as an image). Thus,
if request didn't indicate that the browser was going to treat the
resource as an image (via the 'Sec-Fetch-Dest: image' header), we block
the request when it doesn't return image content.