Temporarily skip Openapi C3 e2e tests
Skipping in yarn due to node version resolution conflicts

The Openapi C3 template depends on `chanfana`, which has a
dependency on `yargs-parser`. The latest chanfana version
(`2.8.1` at the time of this writting) has a dep on
`yargs-parser@22` which requires a node version of `20.19.0`
or higher. Our CI is currently using node version `20.11.1`.
We currently can't bump the CI node version as other tests
will fail, therefore skipping for now until we can properly fix