Ensure container.configuration.image gets written to
Fixes an issue where the user could get into a state where the image has
been pushed but the container fails to update it's configuration. On the
next push container.configuration.image will not be set and
container.image is deleted leading to failures when trying to update the
container app.