Update buildAndMaybePush to use more robust method to find local digests (#9641)
* Update buildAndMaybePush to use more robust method to find local digests

Older versions of docker sometimes will report the digest as "<none>"
which would break with the previous implementation. This implementation
should correctly grab the values.

* Update with PR feedback

* Ensure container.configuration.image gets written to

Fixes an issue where the user could get into a state where the image has
been pushed but the container fails to update it's configuration. On the
next push container.configuration.image will not be set and
container.image is deleted leading to failures when trying to update the
container app.