Add support for Bun WebSockets to miniflare
Bun does not support the upgrade event, but open can be used
since most of this function is just piping responses.

This does not support the use-case of a client reading headers
from the WebSocket response, because Bun does not expose that.

That seems like an edge-case though, so I think merging partial
support is better than none.

See: https://github.com/oven-sh/bun/issues/5951