Support image name without account ID
If a user uses an image in the Cloudflare managed registry they can
specify the image name without including their full account ID. For
example:

  image = "registry.cloudflare.com/hello/world:1.0"

will now "just work", as Wrangler will automatically include their
account ID in the appropriate location when creating the application.
This matches the new behavior in the Dash, where we also omit the
account ID in the image name.