Remove --json arg from cloudchamber and containers commands (#9815)
* Remove --json arg from cloudchamber and containers commands

This is a relic from an earlier time. The only command for which this is
useful is the "cloudchamber curl" subcommand, in which case it should
just be the default and non-configurable anyway.

* Update changeset entry to include omissions for "curl" and "images list"

* Unmock banner to ensure it doesn't print when using --json flag

* Make "cloudchamber apply" always non-interactive

This matches the behavior of "wrangler deploy" which is also always
non-interactive.

* Fix lint errors

* Remove note from changelog about --json with curl command

cloudchamber curl _doesn't_ have a --json flag anymore.

* Rename cloudchamber curl -D option to -d to match real curl

-D is supported as a hidden/deprecated alias