Build containers without account ID (#9765)
* Build container without account ID

This allows us to build containers in dry run mode. When we push the
container to the managed registry, we re-tag it to include the account
ID. We only ever push when not in dry run mode though, so when pushing
we can safely grab the account information.

Co-authored-by: Hasip Timurtaş <hasip.timurtas@gmail.com>

* Always check for Docker when using containers

Even in dry run mode we should check that Docker is installed, because
dry run mode uses docker commands too.

---------

Co-authored-by: Greg Anders <ganders@cloudflare.com>