[R2 Data Catalog] Enable compaction requires a CLOUDFLARE_API_TOKEN
For compaction to work, we need customers to grant our system permission
to access their R2 Bucket and R2 Data Catalog on their behalf. This
permission grant is made via a token they provide when they enable
compaciton. We encrypt the token and store it, and use it only via our
backend system. Because the token is required for compaction to work, we
make it required.