test: only watch config files in ConfigController when not testing
Previously we blocked config watching based on the existence of the `vitest` global. But this made it difficult to override in the tests where we actually needed config file changes to trigger reloads.

Turning on config watching for all tests proved to cause a bunch of tests that don't care about config file changes to become flaky.

So now we are guarding config watching once again but now with the more configurable `process.env.WRANGLER_CI_DISABLE_CONFIG_WATCHING`.