[wrangler] Notify user on local dev server reload (#10750)
Because we set `logRequests: false` in the miniflare options,
miniflare's "Updated and ready" message is suppressed for all
local server reloads.

This change ensures users are notified when the local server is ready
after a reload event, instead of only seeing "Reloading local server..."
with no confirmation of success.

This also changes the `waitForReload` e2e helper function to check
for the reload success message rather than the
"Reloading local server..." message for a stronger guarantee that
reloads have completed.

To accommodate the stronger wait condition, we set the
default timeout of `waitForReload` to 15 seconds to
match the `waitForReady` timeout.